Major changes to layout, settling in on template
Created athelte-homescreen, which is a dialog which contains tabs at
top. The first tab is permanent, but that tab displays a list of
athletes. When a client pushes a row in the list a new tab is created
which displays detailed view of the athlete.

Athlete detail view needs more work, but displays basic info at this
moment.

Added temp route /athlete/home to display the home screen.