More robust User state handling
If the auth guard dos not currently have a user in the authService but
we have a token, we fetch the user info based on that oken from the
backend and then decide if user can access route. Role based auth not
implemented yet, but the preprocess for that has been added.

Added interceptor that checks all responses from backend if it contains
a token, if so we update our token.

Created a model called AuthedUser

Added a route and function that fetches the logged in user data to be
able to update current user information.

Added snackbar for user notification when there was an error logging in.

Created baseUrl in environments

Added actual logged in user to athlete homescreen