add/edit athlete in single component
Both adding and editing athlete is now in a single component. Injected
data of athlete is how the component knows if it's editing old athlete
or adding new. Same validations apply but different functions are called
from authleteService based on if data.athlete was present or not.

Now athleteService stores the athlete list in a BehaviorSubject and an
observable of that list is public to copmonents to subscribe to. This is
to make sure the components can't push data to the stream, only observe.